refactor: Extract shared EvalResult/EvaluationException into evaluator_common
Both evaluators duplicated these pure, I/O-free types verbatim; move them to impl/evaluator_common.py imported by sync and async. Touches the released sync evaluator.py (import + class removal) — behavior identical.
