+++ ldclient/impl/async_evaluator.py
@@ -1,11 +1,12 @@
 import hashlib
 import logging
-from typing import Any, Awaitable, Callable, Dict, List, Optional, Set, Tuple
+from typing import Any, Awaitable, Callable, Optional, Set, Tuple
 
 from ldclient.context import Context
 from ldclient.evaluation import BigSegmentsStatus, EvaluationDetail
 from ldclient.impl import operators
-from ldclient.impl.events.types import EventFactory, EventInputEvaluation
+from ldclient.impl.evaluator_common import EvalResult, EvaluationException
+from ldclient.impl.events.types import EventFactory
 from ldclient.impl.model import *
 
 # For consistency with past logging behavior, we are pretending that the evaluation logic still lives in
@@ -21,50 +22,6 @@ __BUILTINS__ = ["key", "secondary", "ip", "country", "email", "firstName", "last
 # other side effects that are not exposed to the application, such as events generated by
 # prerequisite evaluations, and the cached state of any Big Segments query that we may have
 # ended up having to do for the context.
-class EvalResult:
-    __slots__ = [
-        'detail', 'events', 'big_segments_status', 'big_segments_membership',
-        'original_flag_key', 'prereq_stack', 'segment_stack', 'depth', 'prerequisites'
-    ]
-
-    def __init__(self):
-        self.detail = None
-        self.events = None  # type: Optional[List[EventInputEvaluation]]
-        self.big_segments_status = None  # type: Optional[str]
-        self.big_segments_membership = None  # type: Optional[Dict[str, Optional[dict]]]
-        self.original_flag_key = None  # type: Optional[str]
-        self.prereq_stack = None  # type: Optional[List[str]]
-        self.segment_stack = None  # type: Optional[List[str]]
-        self.depth = 0
-        self.prerequisites = []  # type: List[str]
-
-    def record_prerequisite(self, key: str):
-        if self.depth == 0:
-            self.prerequisites.append(key)
-
-    def add_event(self, event: EventInputEvaluation):
-        if self.events is None:
-            self.events = []
-        self.events.append(event)
-
-    def __repr__(self) -> str:  # used only in test debugging
-        return "EvalResult(detail=%s, events=%s)" % (self.detail, self.events)
-
-
-class EvaluationException(Exception):
-    def __init__(self, message: str, error_kind: str = 'MALFORMED_FLAG'):
-        self._message = message
-        self._error_kind = error_kind
-
-    @property
-    def message(self) -> str:
-        return self._message
-
-    @property
-    def error_kind(self) -> str:
-        return self._error_kind
-
-
 class AsyncEvaluator:
     """
     Async version of Evaluator. Encapsulates the feature flag evaluation logic. The AsyncEvaluator
