+++ ldclient/impl/evaluator_common.py
@@ -1,0 +1,54 @@
+"""
+Shared, I/O-free evaluator types used by both the sync :class:`ldclient.impl.evaluator.Evaluator`
+and the async :class:`ldclient.impl.async_evaluator.AsyncEvaluator`. These contain no store or
+network access, so they are identical across the two evaluators.
+"""
+
+from typing import Dict, List, Optional
+
+from ldclient.impl.events.types import EventInputEvaluation
+
+
+# EvalResult is used internally to hold the EvaluationDetail result of an evaluation along with
+# other side effects that are not exposed to the application, such as events generated by
+# prerequisite evaluations, and the cached state of any Big Segments query that we may have
+# ended up having to do for the context.
+class EvalResult:
+    __slots__ = ['detail', 'events', 'big_segments_status', 'big_segments_membership', 'original_flag_key', 'prereq_stack', 'segment_stack', 'depth', 'prerequisites']
+
+    def __init__(self):
+        self.detail = None
+        self.events = None  # type: Optional[List[EventInputEvaluation]]
+        self.big_segments_status = None  # type: Optional[str]
+        self.big_segments_membership = None  # type: Optional[Dict[str, Optional[dict]]]
+        self.original_flag_key = None  # type: Optional[str]
+        self.prereq_stack = None  # type: Optional[List[str]]
+        self.segment_stack = None  # type: Optional[List[str]]
+        self.depth = 0
+        self.prerequisites = []  # type: List[str]
+
+    def record_prerequisite(self, key: str):
+        if self.depth == 0:
+            self.prerequisites.append(key)
+
+    def add_event(self, event: EventInputEvaluation):
+        if self.events is None:
+            self.events = []
+        self.events.append(event)
+
+    def __repr__(self) -> str:  # used only in test debugging
+        return "EvalResult(detail=%s, events=%s)" % (self.detail, self.events)
+
+
+class EvaluationException(Exception):
+    def __init__(self, message: str, error_kind: str = 'MALFORMED_FLAG'):
+        self._message = message
+        self._error_kind = error_kind
+
+    @property
+    def message(self) -> str:
+        return self._message
+
+    @property
+    def error_kind(self) -> str:
+        return self._error_kind
